Implement a pm.System shorcut
Since lots of modules need to execute some command at somepoint
in sync manner (waiting for result) i thought it's handy and cleaner
to introduce a pm.System short cut so not each module try to do it
in its own way.

For async mode, or to make use of advanced features like hooks, modules
will still to do it the usual way.